HH-118342 move StatusServlet to sepatate file
