+++ nab-starter/src/main/java/ru/hh/nab/starter/servlet/StatusServletConfig.java
@@ -1,0 +1,27 @@
+package ru.hh.nab.starter.servlet;
+
+import javax.servlet.Servlet;
+import org.glassfish.jersey.server.ResourceConfig;
+import org.glassfish.jersey.servlet.ServletContainer;
+import org.springframework.web.context.WebApplicationContext;
+import ru.hh.nab.starter.AppMetadata;
+import ru.hh.nab.starter.resource.StatusResource;
+
+public class StatusServletConfig implements NabServletConfig {
+  @Override
+  public String[] getMapping() {
+    return new String[]{"/status"};
+  }
+
+  @Override
+  public String getName() {
+    return "status";
+  }
+
+  @Override
+  public Servlet createServlet(WebApplicationContext rootCtx) {
+    ResourceConfig statusResourceConfig = new ResourceConfig();
+    statusResourceConfig.register(new StatusResource(rootCtx.getBean(AppMetadata.class)));
+    return new ServletContainer(statusResourceConfig);
+  }
+}
